requirements.txt
python-binance>=1.0.17
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
ccxt>=3.0.27
ta>=0.10.2
pydantic>=2.0.0
loguru>=0.7.0
schedule>=1.2.0
Flask>=2.3.0
